Upgrade dependency
Google's disruptive updates are too frequent and unregulated.
If you modify the way the old API is used, then you should declare it as the new API.

After the update, neither the compiler nor the IDE reported an error. Not until debugging. And there is no documentation on how to change it, I learned about the new usage by looking at the official sample. This wasted a lot of my time.